requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pyrogram==2.0.106
celery==5.3.6
django-apscheduler==0.6.2
APScheduler==3.10.4
django-cors-headers==4.2.0
uvicorn==0.14.0
openai==1.14.3
pillow==10.3.0
